move the AMDS PV control related codes to the folder PVController
