+++ AMDS_sgm_pv.pro
@@ -18,8 +18,8 @@ HEADERS *= \
 	source/application/AMDSCentralServer.h \
         source/application/SGM/AMDSCentralServerSGMPV.h \
         source/Detector/AMDSDetectorServer.h \
-    source/Detector/AMDSPVConfigurationMap.h \
-    source/Detector/AMDSPVController.h
+        source/Detector/PVController/AMDSPVConfigurationMap.h \
+        source/Detector/PVController/AMDSPVController.h
 #	source/Detector/Scaler/AMDSScalerCommandManager.h \
 #	source/Detector/Scaler/AMDSScalerDetector.h \
 #	source/Detector/Scaler/AMDSScalerDetectorServer.h \
@@ -30,7 +30,7 @@ SOURCES *= \
 	source/application/AMDSCentralServer.cpp \
         source/application/SGM/AMDSCentralServerSGMPV.cpp \
         source/Detector/AMDSDetectorServer.cpp \
-    source/Detector/AMDSPVController.cpp
+        source/Detector/PVController/AMDSPVController.cpp
 #	source/Detector/Scaler/AMDSScalerCommandManager.cpp \
 #	source/Detector/Scaler/AMDSScalerDetector.cpp \
 #	source/Detector/Scaler/AMDSScalerDetectorServer.cpp \
