+++ source/application/SGM/AMDSCentralServerSGMPV.h
@@ -9,9 +9,6 @@
 class AMDSPVConfigurationMap;
 class AMDSPVControllerManager;
 
-//class AMDSScalerDetectorManager;
-//class AMDSDetectorServerManager;
-
 #define AMDS_SGM_SERVER_ALT_INVALID_BUFFERGROUP_NAME 30101
 
 class AMDSCentralServerSGMPV : public AMDSCentralServer
